+++ packages/blurry-image-lazy-loading/src/api/api-extensions.ts
@@ -15,6 +15,7 @@ export const adminApiExtensions = gql`
   type PluginPreviewImageHashResult {
     code: PluginPreviewImageHashResultCode!
     jobsAddedToQueue: Int!
+    assetsSkipped: Int!
     message: String!
   }
 
@@ -65,6 +66,15 @@ export const adminApiExtensions = gql`
     @default true
     """
     deduplicateAssetIds: Boolean
+
+    """
+    Collections can be quite large, so by default this mutation skips over assets that already have a hash set, in order to minimize the needed compute.
+    This can be undesirable, for example after you change the strategy, the encoding or resize options.
+    Setting this option to \`true\` will overwrite the existing hashes, letting you update your existing assets.
+
+    @default false
+    """
+    regenerateExistingHashes: Boolean
   }
 
   input PluginPreviewImageHashForProductInput {
@@ -87,6 +97,13 @@ export const adminApiExtensions = gql`
     """
     Create preview image hashes for an entire collection.
     This includes the collection, the contained Product-assets and related ProductVariant-assets.
+
+    Due to how large collections can become, you may want to disable the deduplication of asset ids.
+    If deduplication is enabled, jobs will be created only after gathering all assets first.
+    If disabled, jobs will be created as the assets are being read.
+
+    No deduplication may result in assets being hashed multiple times, but the tradeoff is not having
+    to hold potentially millions of records in memory and just letting the worker take care of them eventually.
     """
     pluginPreviewImageHashCreateImageHashesForCollection(
       input: PluginPreviewImageHashForCollectionInput!